moved base.whitelist_method_generator to .generic
