+++ pandas/_typing.py
@@ -9,7 +9,7 @@ from pandas._libs.tslibs.timedeltas import Timedelta
 
 from pandas.core.dtypes.dtypes import ExtensionDtype
 from pandas.core.dtypes.generic import (
-    ABCExtensionArray, ABCIndexClass, ABCSeries, ABCSparseSeries)
+    ABCDataFrame, ABCExtensionArray, ABCIndexClass, ABCSeries, ABCSparseSeries)
 
 AnyArrayLike = TypeVar('AnyArrayLike',
                        ABCExtensionArray,
@@ -22,3 +22,5 @@ DatetimeLikeScalar = TypeVar('DatetimeLikeScalar', Period, Timestamp,
                              Timedelta)
 Dtype = Union[str, np.dtype, ExtensionDtype]
 FilePathOrBuffer = Union[str, Path, IO[AnyStr]]
+
+FrameOrSeries = TypeVar('FrameOrSeries', ABCSeries, ABCDataFrame)
